+++ tests/de/halirutan/mathematica/refactoring/RenameTest.java
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2016 Patrick Scheibe
+ * Copyright (c) 2018 Patrick Scheibe
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
@@ -7,16 +8,16 @@
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
  *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
  * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.halirutan.mathematica.refactoring;
@@ -24,9 +25,6 @@ package de.halirutan.mathematica.refactoring;
 import com.intellij.testFramework.fixtures.LightCodeInsightFixtureTestCase;
 import de.halirutan.mathematica.file.MathematicaFileType;
 
-/**
- * @author patrick (24.12.16).
- */
 public class RenameTest extends LightCodeInsightFixtureTestCase{
 
   public void testNormalVariable() {
@@ -77,8 +75,7 @@ public class RenameTest extends LightCodeInsightFixtureTestCase{
 
   //TODO: This seems to be a bug in the FindUsage framework of IDEA
   //I have to decide if I really want to do this work by myself and probably loose performance.
-/*
-  public void testDollarVariables() throws Exception {
+  public void testDollarVariables() {
     myFixture.configureByText(MathematicaFileType.INSTANCE,
         "var$ = 1;\n" +
             "var$<caret> + var$");
@@ -86,6 +83,5 @@ public class RenameTest extends LightCodeInsightFixtureTestCase{
     myFixture.checkResult("$var$ = 1;\n" +
         "$var$ + $var$");
   }
-*/
 
 }